python_services/terrain_generation/requirements.txt
black==24.10.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
geopandas==1.0.1
grpcio==1.67.0
grpcio-tools==1.67.0
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
mypy-extensions==1.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.28.2
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.2.0
shapely==2.0.6
six==1.16.0
tzdata==2024.2

